# perf: optimied map iteration for scalar types