--- a/internal/reflect/append_list_gen.go
+++ b/internal/reflect/append_list_gen.go
@@ -22,20 +22,20 @@ import "unsafe"
 // Template and code can be found in append_list_gen_test.go.
 
 func init() {
-	registerListAppendFunc(tBYTE, appendList_I08)
-	registerListAppendFunc(tI16, appendList_I16)
-	registerListAppendFunc(tI32, appendList_I32)
-	registerListAppendFunc(tI64, appendList_I64)
-	registerListAppendFunc(tDOUBLE, appendList_I64)
-	registerListAppendFunc(tENUM, appendList_ENUM)
-	registerListAppendFunc(tSTRING, appendList_STRING)
-	registerListAppendFunc(tSTRUCT, appendList_Other)
-	registerListAppendFunc(tMAP, appendList_Other)
-	registerListAppendFunc(tSET, appendList_Other)
-	registerListAppendFunc(tLIST, appendList_Other)
+	registerListAppendFunc(tI08, appendList_tI08)
+	registerListAppendFunc(tI16, appendList_tI16)
+	registerListAppendFunc(tI32, appendList_tI32)
+	registerListAppendFunc(tI64, appendList_tI64)
+	registerListAppendFunc(tDOUBLE, appendList_tI64)
+	registerListAppendFunc(tENUM, appendList_tENUM)
+	registerListAppendFunc(tSTRING, appendList_tSTRING)
+	registerListAppendFunc(tSTRUCT, appendList_tOTHER)
+	registerListAppendFunc(tMAP, appendList_tOTHER)
+	registerListAppendFunc(tSET, appendList_tOTHER)
+	registerListAppendFunc(tLIST, appendList_tOTHER)
 }
 
-func appendList_I08(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tI08(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
@@ -45,12 +45,12 @@ func appendList_I08(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		b = append(b, *((*byte)(vp)))
+		b = append(b, *(*byte)(vp))
 	}
 	return b, nil
 }
 
-func appendList_I16(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tI16(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
@@ -60,12 +60,12 @@ func appendList_I16(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		b = appendUint16(b, *((*uint16)(vp)))
+		b = appendUint16(b, *(*uint16)(vp))
 	}
 	return b, nil
 }
 
-func appendList_I32(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tI32(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
@@ -75,12 +75,12 @@ func appendList_I32(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		b = appendUint32(b, *((*uint32)(vp)))
+		b = appendUint32(b, *(*uint32)(vp))
 	}
 	return b, nil
 }
 
-func appendList_I64(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tI64(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
@@ -90,12 +90,12 @@ func appendList_I64(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		b = appendUint64(b, *((*uint64)(vp)))
+		b = appendUint64(b, *(*uint64)(vp))
 	}
 	return b, nil
 }
 
-func appendList_ENUM(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tENUM(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
@@ -105,40 +105,39 @@ func appendList_ENUM(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		b = appendUint32(b, uint32(*((*int64)(vp))))
+		b = appendUint32(b, uint32(*(*uint64)(vp)))
 	}
 	return b, nil
 }
 
-func appendList_STRING(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tSTRING(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
 		return b, nil
 	}
-	var s string
 	for i := uint32(0); i < n; i++ {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
-		s = *((*string)(vp))
+		s := *(*string)(vp)
 		b = appendUint32(b, uint32(len(s)))
 		b = append(b, s...)
 	}
 	return b, nil
 }
 
-func appendList_Other(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
+func appendList_tOTHER(t *tType, b []byte, p unsafe.Pointer) ([]byte, error) {
 	t = t.V
 	b, n, vp := appendListHeader(t, b, p)
 	if n == 0 {
 		return b, nil
 	}
-	var err error
 	for i := uint32(0); i < n; i++ {
 		if i != 0 {
 			vp = unsafe.Add(vp, t.Size)
 		}
+		var err error
 		if t.IsPointer {
 			b, err = t.AppendFunc(t, b, *(*unsafe.Pointer)(vp))
 		} else {