--- a/internal/reflect/decoder.go
+++ b/internal/reflect/decoder.go
@@ -19,11 +19,9 @@ package reflect
 import (
 	"encoding/binary"
 	"fmt"
-	"reflect"
 	"sync"
 	"unsafe"
 
-	"github.com/cloudwego/frugal/internal/defs"
 	"github.com/cloudwego/gopkg/protocol/thrift"
 )
 
@@ -104,28 +102,31 @@ func (d *tDecoder) Decode(b []byte, base unsafe.Pointer, sd *structDesc, maxdept
 			i += n
 			continue
 		}
+		if bs != nil { // update bitset if has required fields
+			bs.set(f.ID)
+		}
 		p := unsafe.Add(base, f.Offset) // pointer to the field
 
 		t := f.Type
 		p = d.mallocIfPointer(t, p)
 		if t.FixedSize > 0 {
+			// fixed len types, can be inlined without func call
 			i += decodeFixedSizeTypes(t.T, b[i:], p)
+			continue
+		}
+
+		var n int
+		var err error
+		if f.NoCopy {
+			n, err = decodeStringNoCopy(d, t, b[i:], p)
 		} else {
-			var n int
-			var err error
-			if f.NoCopy {
-				n, err = decodeStringNoCopy(t, b[i:], p)
-			} else {
-				n, err = d.decodeType(t, b[i:], p, maxdepth-1)
-			}
-			if err != nil {
-				return i, fmt.Errorf("decode field %d of struct %s err: %w", fid, sd.rt.String(), err)
-			}
-			i += n
+			n, err = t.DecodeFunc(d, t, b[i:], p, maxdepth-1)
 		}
-		if bs != nil {
-			bs.set(f.ID)
+		if err != nil {
+			return i, fmt.Errorf("decode field %d of struct %s err: %w", fid, sd.rt.String(), err)
 		}
+		i += n
+
 	}
 	for _, fid := range sd.requiredFieldIDs {
 		if !bs.test(fid) {
@@ -138,254 +139,11 @@ func (d *tDecoder) Decode(b []byte, base unsafe.Pointer, sd *structDesc, maxdept
 	return i, nil
 }
 
-func decodeFixedSizeTypes(t ttype, b []byte, p unsafe.Pointer) int {
-	switch t {
-	case tBOOL, tBYTE:
-		*(*byte)(p) = b[0] // XXX: for tBOOL 1->true, 2->true/false
-		return 1
-	case tDOUBLE, tI64:
-		*(*uint64)(p) = binary.BigEndian.Uint64(b)
-		return 8
-	case tI16:
-		*(*int16)(p) = int16(binary.BigEndian.Uint16(b))
-		return 2
-	case tI32:
-		*(*int32)(p) = int32(binary.BigEndian.Uint32(b))
-		return 4
-	case tENUM:
-		*(*int64)(p) = int64(int32(binary.BigEndian.Uint32(b)))
-		return 4
-	default:
-		panic("bug")
+func decodeStruct(d *tDecoder, t *tType, b []byte, p unsafe.Pointer, maxdepth int) (int, error) {
+	if t.Sd.hasInitFunc {
+		f := t.Sd.initFunc // copy on write, reuse itab of iface
+		updateIface(unsafe.Pointer(&f), p)
+		f.InitDefault()
 	}
-}
-
-func decodeStringNoCopy(t *tType, b []byte, p unsafe.Pointer) (i int, err error) {
-	l := int(binary.BigEndian.Uint32(b))
-	if l < 0 {
-		err = errNegativeSize
-		return
-	}
-	i += 4
-	if l == 0 {
-		if t.Tag == defs.T_binary {
-			*(*[]byte)(p) = []byte{}
-		} else {
-			*(*string)(p) = ""
-		}
-		return
-	}
-
-	// assert len, panic if []byte shorter than expected.
-	_ = b[i+l-1]
-
-	if t.Tag == defs.T_binary {
-		h := (*sliceHeader)(p)
-		h.Data = uintptr(unsafe.Pointer(&b[i]))
-		h.Len = l
-		h.Cap = l
-	} else { //  convert to str
-		h := (*stringHeader)(p)
-		h.Data = uintptr(unsafe.Pointer(&b[i]))
-		h.Len = l
-	}
-	i += l
-	return
-}
-
-func (d *tDecoder) decodeType(t *tType, b []byte, p unsafe.Pointer, maxdepth int) (int, error) {
-	if maxdepth == 0 {
-		return 0, errDepthLimitExceeded
-	}
-	if t.FixedSize > 0 {
-		return decodeFixedSizeTypes(t.T, b, p), nil
-	}
-	switch t.T {
-	case tSTRING:
-		l := int(binary.BigEndian.Uint32(b))
-		if l < 0 {
-			return 0, errNegativeSize
-		}
-		i := 4
-		if l == 0 {
-			if t.Tag == defs.T_binary {
-				*(*[]byte)(p) = []byte{}
-			} else {
-				*(*string)(p) = ""
-			}
-			return i, nil
-		}
-
-		// assert len, panic if []byte shorter than expected.
-		_ = b[i+l-1]
-
-		x := d.Malloc(l, 1, 0)
-		if t.Tag == defs.T_binary {
-			h := (*sliceHeader)(p)
-			h.Data = uintptr(x)
-			h.Len = l
-			h.Cap = l
-		} else { //  convert to str
-			h := (*stringHeader)(p)
-			h.Data = uintptr(x)
-			h.Len = l
-		}
-		copyn(x, b[i:], l)
-		i += l
-		return i, nil
-	case tMAP:
-		// map header
-		t0, t1, l := ttype(b[0]), ttype(b[1]), int(binary.BigEndian.Uint32(b[2:]))
-		if l < 0 {
-			return 0, errNegativeSize
-		}
-
-		// check types
-		kt := t.K
-		vt := t.V
-		if t0 != kt.WT || t1 != vt.WT {
-			return 0, newTypeMismatchKV(kt.WT, vt.WT, t0, t1)
-		}
-
-		// decode map
-
-		// tmp vars
-		// tmpk = decode(b)
-		// tmpv = decode(b)
-		// map[tmpk] = tmpv
-		tmp := t.MapTmpVarsPool.Get().(*tmpMapVars)
-		k := tmp.k
-		v := tmp.v
-		kp := tmp.kp
-		vp := tmp.vp
-		m := reflect.MakeMapWithSize(t.RT, l)
-		*((*uintptr)(p)) = m.Pointer() // p = make(t.RT, l)
-
-		// pre-allocate space for keys and values if they're pointers
-		// like
-		// kp = &sliceK[i], decode(b, kp)
-		// instead of
-		// kp = new(type), decode(b, kp)
-		var sliceK unsafe.Pointer
-		if kt.IsPointer && l > 0 {
-			sliceK = d.Malloc(l*kt.V.Size, kt.V.Align, kt.V.MallocAbiType)
-		}
-		var sliceV unsafe.Pointer
-		if vt.IsPointer && l > 0 {
-			sliceV = d.Malloc(l*vt.V.Size, vt.V.Align, vt.V.MallocAbiType)
-		}
-
-		var n int
-		var err error
-		i := 6
-		for j := 0; j < l; j++ {
-			p = kp
-			if kt.IsPointer { // p = &sliceK[j]
-				if j != 0 {
-					sliceK = unsafe.Add(sliceK, kt.V.Size) // next
-				}
-				*(*unsafe.Pointer)(p) = sliceK
-				p = sliceK
-			}
-			if kt.FixedSize > 0 {
-				i += decodeFixedSizeTypes(kt.T, b[i:], p)
-			} else {
-				if n, err = d.decodeType(kt, b[i:], p, maxdepth-1); err != nil {
-					break
-				} else {
-					i += n
-				}
-			}
-			p = vp
-			if vt.IsPointer { // p = &sliceV[j]
-				if j != 0 { // next
-					sliceV = unsafe.Add(sliceV, vt.V.Size)
-				}
-				*(*unsafe.Pointer)(p) = sliceV
-				p = sliceV
-			}
-			if vt.FixedSize > 0 {
-				i += decodeFixedSizeTypes(vt.T, b[i:], p)
-			} else {
-				if n, err = d.decodeType(vt, b[i:], p, maxdepth-1); err != nil {
-					break
-				} else {
-					i += n
-				}
-			}
-			m.SetMapIndex(k, v)
-		}
-		t.MapTmpVarsPool.Put(tmp) // no defer, it may be in hot path
-		return i, nil
-	case tLIST, tSET: // NOTE: for tSET, it may be map in the future
-		// list header
-		tp, l := ttype(b[0]), int(binary.BigEndian.Uint32(b[1:]))
-		if l < 0 {
-			return 0, errNegativeSize
-		}
-		// check types
-		et := t.V
-		if et.WT != tp {
-			return 0, newTypeMismatch(et.WT, tp)
-		}
-
-		i := 5
-
-		// decode list
-		h := (*sliceHeader)(p) // update the slice field
-		if l <= 0 {
-			h.Zero()
-			return i, nil
-		}
-		x := d.Malloc(l*et.Size, et.Align, et.MallocAbiType) // malloc for slice. make([]Type, l, l)
-		h.Data = uintptr(x)
-		h.Len = l
-		h.Cap = l
-
-		// pre-allocate space for elements if they're pointers
-		// like
-		// v[i] = &sliceData[i]
-		// instead of
-		// v[i] = new(type)
-		var sliceData unsafe.Pointer
-		if et.IsPointer {
-			sliceData = d.Malloc(l*et.V.Size, et.V.Align, et.V.MallocAbiType)
-		}
-
-		p = x // point to the 1st element, and then decode one by one
-		for j := 0; j < l; j++ {
-			if j != 0 {
-				p = unsafe.Add(p, et.Size) // next element
-			}
-			vp := p // v[j]
-
-			// p = &sliceData[j], see comment of sliceData above
-			if et.IsPointer {
-				if j != 0 {
-					sliceData = unsafe.Add(sliceData, et.V.Size) // next
-				}
-				*(*unsafe.Pointer)(p) = sliceData // v[j] = &sliceData[i]
-				vp = sliceData                    // &v[j]
-			}
-
-			if et.FixedSize > 0 {
-				i += decodeFixedSizeTypes(et.T, b[i:], vp)
-			} else {
-				n, err := d.decodeType(et, b[i:], vp, maxdepth-1)
-				if err != nil {
-					return i, err
-				}
-				i += n
-			}
-		}
-		return i, nil
-	case tSTRUCT:
-		if t.Sd.hasInitFunc {
-			f := t.Sd.initFunc // copy on write, reuse itab of iface
-			updateIface(unsafe.Pointer(&f), p)
-			f.InitDefault()
-		}
-		return d.Decode(b, p, t.Sd, maxdepth-1)
-	}
-	return 0, fmt.Errorf("unknown type: %d", t.T)
+	return d.Decode(b, p, t.Sd, maxdepth)
 }