--- a/internal/reflect/decoder_gen.sh
+++ b/internal/reflect/decoder_gen.sh
@@ -1,0 +1,21 @@
+#!/bin/bash
+#
+# Copyright 2025 CloudWeGo Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+FRUGAL_GEN_DECODE_MAP_FILE="decoder_map_gen.go"
+
+rm -f $FRUGAL_GEN_DECODE_MAP_FILE
+
+exec go test -v -run=TestGenDecoder -gencode=true