… internal/reflect/append_list_gen_test.go
… internal/reflect/append_map_gen.go
… internal/reflect/append_map_gen_test.go
… internal/reflect/append_map_test.go
… internal/reflect/decoder_list_test.go
… internal/reflect/decoder_map_gen.go
… internal/reflect/decoder_map_gen_test.go
… internal/reflect/decoder_map_test.go
… internal/reflect/decoder_test.go