goos: linux
goarch: amd64
pkg: github.com/cloudwego/frugal/tests
cpu: Intel(R) Xeon(R) Gold 5118 CPU @ 2.30GHz
                                  │  ./old.txt   │              new.txt               │
                                  │    sec/op    │   sec/op     vs base               │
AllSize_Marshal_Frugal/small-4       92.63n ± 4%   92.15n ± 6%        ~ (p=0.097 n=7)
AllSize_Marshal_Frugal/medium-4      3.692µ ± 0%   3.504µ ± 1%   -5.09% (p=0.001 n=7)
AllSize_Marshal_Frugal/large-4       64.39µ ± 1%   60.72µ ± 0%   -5.69% (p=0.001 n=7)
AllSize_Unmarshal_Frugal/small-4     237.6n ± 4%   240.7n ± 1%        ~ (p=0.302 n=7)
AllSize_Unmarshal_Frugal/medium-4   10.145µ ± 1%   8.847µ ± 8%  -12.79% (p=0.001 n=7)
AllSize_Unmarshal_Frugal/large-4     179.9µ ± 1%   156.0µ ± 3%  -13.29% (p=0.001 n=7)
geomean                              4.606µ        4.321µ        -6.18%

                                  │  ./old.txt   │               new.txt               │
                                  │     B/s      │     B/s       vs base               │
AllSize_Marshal_Frugal/small-4      5.862Gi ± 4%   5.892Gi ± 6%        ~ (p=0.097 n=7)
AllSize_Marshal_Frugal/medium-4     4.419Gi ± 0%   4.656Gi ± 1%   +5.35% (p=0.001 n=7)
AllSize_Marshal_Frugal/large-4      4.392Gi ± 1%   4.657Gi ± 0%   +6.04% (p=0.001 n=7)
AllSize_Unmarshal_Frugal/small-4    2.285Gi ± 4%   2.255Gi ± 1%        ~ (p=0.318 n=7)
AllSize_Unmarshal_Frugal/medium-4   1.608Gi ± 1%   1.844Gi ± 8%  +14.68% (p=0.001 n=7)
AllSize_Unmarshal_Frugal/large-4    1.572Gi ± 1%   1.813Gi ± 3%  +15.33% (p=0.001 n=7)
geomean                             2.949Gi        3.143Gi        +6.58%

#### What type of PR is this?
<!--
Add one of the following kinds:

build: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
ci: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
docs: Documentation only changes
feat: A new feature
optimize: A new optimization
fix: A bug fix
perf: A code change that improves performance
refactor: A code change that neither fixes a bug nor adds a feature
style: Changes that do not affect the meaning of the code (white space, formatting, missing semi-colons, etc)
test: Adding missing tests or correcting existing tests
chore: Changes to the build process or auxiliary tools and libraries such as documentation generation
-->

#### Check the PR title.
perf

<!--
The description of the title will be attached in Release Notes, 
so please describe it from user-oriented, what this PR does / why we need it.
Please check your PR title with the below requirements:
-->
- [ ] This PR title match the format: \<type\>(optional scope): \<description\>
- [ ] The description of this PR title is user-oriented and clear enough for others to understand.
- [ ] Attach the PR updating the user documentation if the current PR requires user awareness at the usage level. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)


#### (Optional) Translate the PR title into Chinese.


#### (Optional) More detailed description for this PR(en: English/zh: Chinese).
<!--
Provide more detailed info for review(e.g., it's recommended to provide perf data if this is a perf type PR).
-->
en:
zh(optional): 


#### (Optional) Which issue(s) this PR fixes:
<!--
Automatically closes linked issue when PR is merged.
Eg: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

#### (optional) The PR that updates user documentation:
<!--
If the current PR requires user awareness at the usage level, please submit a PR to update user docs. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)
-->